Move FromString method to a separate file: Fraction.ConvertFromString.cs (was a suggestion from @lipchev in a previous PR - which simplifies code reviews).
